More embed stuff

+++ src/main/kotlin/event/EventHandler.kt
@@ -3,10 +3,13 @@ package event
 import BOT
 import CONFIG
 import LOGGER
+import net.dv8tion.jda.api.entities.channel.concrete.TextChannel
+import net.dv8tion.jda.api.entities.emoji.Emoji
 import net.dv8tion.jda.api.events.ReadyEvent
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
+import net.dv8tion.jda.internal.requests.Route.Emojis
 import util.CommandUtil
 import util.EmbedUtil
 
@@ -17,11 +20,14 @@ class EventHandler: ListenerAdapter() {
     }
 
     override fun onMessageReceived(event: MessageReceivedEvent) {
-        if (event.channel.idLong == CONFIG.suggestionsId) {
+        if (event.channel.idLong == CONFIG.suggestionsId && event.message.author != BOT.jda.selfUser) {
             val message = event.message
-            message.createThreadChannel(message.contentRaw)
+            message.createThreadChannel(message.contentRaw).queue()
+            message.addReaction(Emoji.fromUnicode("\uD83D\uDC4D")).queue()
+            message.addReaction(Emoji.fromUnicode("\uD83D\uDC4E")).queue()
             event.channel.iterableHistory.find { it.author == BOT.jda.selfUser }?.delete()?.queue()
-            // event.channel.sendMessageEmbeds(EmbedUtil)
+            val (embed, _) = EmbedUtil.customEmbed("suggestions")
+            event.channel.sendMessageEmbeds(embed).queue()
         }
     }
